+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # batoanng/eslint-config
 
+## 3.2.0
+
+- Add new config, synchronize versions
+
 ## 3.1.0
 
 ### Minor Changes