+++ packages/jest-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @batoanng/jest-config
 
+## 1.3.0
+
+- Add new config, synchronize versions
+
 ## 1.2.0
 
 ### Minor Changes