+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @batoanng/utils
 
+## 2.3.0
+
+- Add new config, synchronize versions
+
 ## 2.2.0
 
 ### Minor Changes